% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FileIO.R
\name{addPostfix}
\alias{addPostfix}
\title{Returns filename, with postfixes appended, each separated by dots, followed by the filename extension}
\usage{
addPostfix(filename, ...)
}
\arguments{
\item{...}{(character) Filename to append postfixes to}

\item{...}{(character) Postfixes to append to filename}
}
\value{
(character) postfixed filename
}
\description{
Returns filename, with postfixes appended, each separated by dots, followed by the filename extension
}
\examples{
# Returns "data.a.b.c.csv"
addPostfix( "data.csv", "a", "b", "c" );
}

