% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateScores.R
\name{calculateAggregation}
\alias{calculateAggregation}
\title{Calculates scores from trial data using the 'aggregation' approach}
\usage{
calculateAggregation(ds, settings, splithalves = 0,
  splithalf_method = "pearson", verbose = F, ...)
}
\arguments{
\item{ds}{(data.frame) Trial data.}

\item{settings}{(list) D-score calculation settings.}

\item{splithalves}{(integer) If 0, return a single set of d-scores, if > 0, return randomized split-halve reliability averaged over splithalves iterations.}

\item{verbose}{(logical) If TRUE, then print debug output.}

\item{...}{Remaining arguments are passed to an internal scoring function. Not used at the moment.}
}
\value{
(data.frame) Calculated d-scores.
}
\description{
Calculates scores from trial data using the 'aggregation' approach
}
\details{
The table below shows an overview of elements in settings and their meaning.
\tabular{lllll}{
  \strong{Name} \tab \strong{Type} \tab \strong{Required?} \tab  \strong{Example Value} \tab \strong{Description} \cr
  aggregation_factors \tab character \tab Yes \tab \code{} \tab Apply aggregation_factor_function to each subset of ds formed by unique combinations of values for these variables \cr
  aggregation_factor_function \tab character \tab Yes \tab \code{} \tab Function applied to calculate aggregated scores (e.g. 'median rt for correct responses') \cr
  aggregation_run_function \tab character \tab Yes \tab \code{} \tab Function applied to output of aggregation_factor_function (e.g. difference of medians between red and blue conditions) \cr
}
}
\examples{
See: SANDRA/framework_demos/scripts/0.t.2 Calculate Scores.R
}
\seealso{
\code{\link[sandra]{dropAndReport}} for additional settings that can be used for 
dropping and reporting responses and response times before scores are calculated.

Other SANDRA: \code{\link{EasyFit}}, \code{\link{FileIO}},
  \code{\link{FrameworkFileIO}},
  \code{\link{ProgressTimer}}, \code{\link{addPostfix}},
  \code{\link{appendData}}, \code{\link{calculateDScores}},
  \code{\link{calculateScores}},
  \code{\link{checkAndRemoveJasminDuplicates}},
  \code{\link{cutExtension}},
  \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.new}},
  \code{\link{data.frame.renameVar}},
  \code{\link{decodeJasmin1}},
  \code{\link{decodeJasmin2Legacy}},
  \code{\link{decodeJasmin2}}, \code{\link{decodeSprif1}},
  \code{\link{done}}, \code{\link{dropAndReport}},
  \code{\link{existsData}}, \code{\link{getExtension}},
  \code{\link{installAnalysisFramework}},
  \code{\link{leftMerge}}, \code{\link{makeWide}},
  \code{\link{niceBy}}, \code{\link{niceMerge}},
  \code{\link{path}}, \code{\link{readData}},
  \code{\link{readSurveyTool}}, \code{\link{readVector}},
  \code{\link{reportProgress}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{timeRemaining}},
  \code{\link{writeData}}, \code{\link{writeVector}}
}

