% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateDScores.R
\name{calculateDScores}
\alias{calculateDScores}
\title{Calculates d-scores from trial data}
\usage{
calculateDScores(ds, settings, splithalves = 0, what = "all", verbose = F,
  ...)
}
\arguments{
\item{ds}{(data.frame) Trial data}

\item{settings}{(list) D-score calculation settings}

\item{splithalves}{(integer) If 0, return a single set of d-scores, if > 0, return randomized split-halve reliability averaged over splithalves iterations}

\item{what}{(character) If "all", calculate over full task data, if "first_second" calculate 2 dscores; one for first and one for second halve. Only efffectie if splithalves == 0}

\item{verbose}{(logical) If TRUE, then print debug output}
}
\value{
(data.frame) Calculated d-scores
}
\description{
Calculates d-scores from trial data
}
\examples{
See: SANDRA/tests/scripts/demo.jasmin1_data.2.calculateDScores.R
}

