% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAndRemoveJasminDuplicates.R
\name{checkAndRemoveJasminDuplicates}
\alias{checkAndRemoveJasminDuplicates}
\title{Checks JASMIN1 and JASMIN2 data for duplicates}
\usage{
checkAndRemoveJasminDuplicates(evlogs, snVar = "logger_sn",
  timeVar = "logger_time")
}
\arguments{
\item{evlogs}{(data.frame) JASMIN1 or JASMIN2 trial data}
}
\value{
(list) Element "evlogs" contains deduplicated data; Element "sequence_report" contains a report about whether anything suspicious was found. See "2. Decoding Trial Data, Scoring Tasks, and Widening" for information about the sequence_report
}
\description{
Checks JASMIN1 and JASMIN2 data for duplicates
}
\seealso{
Other SANDRA: \code{\link{EasyFit}}, \code{\link{FileIO}},
  \code{\link{FrameworkFileIO}},
  \code{\link{ProgressTimer}}, \code{\link{addPostfix}},
  \code{\link{appendData}},
  \code{\link{calculateAggregation}},
  \code{\link{calculateDScores}},
  \code{\link{calculateScores}},
  \code{\link{cutExtension}},
  \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.new}},
  \code{\link{data.frame.renameVar}},
  \code{\link{decodeJasmin1}},
  \code{\link{decodeJasmin2Legacy}},
  \code{\link{decodeJasmin2}}, \code{\link{decodeSprif1}},
  \code{\link{done}}, \code{\link{dropAndReport}},
  \code{\link{existsData}}, \code{\link{getExtension}},
  \code{\link{installAnalysisFramework}},
  \code{\link{leftMerge}}, \code{\link{makeWide}},
  \code{\link{niceBy}}, \code{\link{niceMerge}},
  \code{\link{path}}, \code{\link{readData}},
  \code{\link{readSurveyTool}}, \code{\link{readVector}},
  \code{\link{reportProgress}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{timeRemaining}},
  \code{\link{updateChecksum}}, \code{\link{writeData}},
  \code{\link{writeVector}}
}

