% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProgressTimer.R
\name{close}
\alias{close}
\title{Close progress bar}
\usage{
close()
}
\description{
Close progress bar
}
\seealso{
Other SANDRA: \code{\link{EasyLM}}, \code{\link{FileIO}},
  \code{\link{FrameworkFileIO}},
  \code{\link{ProgressTimer}}, \code{\link{addPostfix}},
  \code{\link{appendData}}, \code{\link{calculateDScores}},
  \code{\link{cutExtension}},
  \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.new}},
  \code{\link{data.frame.renameVar}},
  \code{\link{decodeJasmin1}}, \code{\link{existsData}},
  \code{\link{getExtension}},
  \code{\link{installAnalysisFramework}},
  \code{\link{leftMerge}}, \code{\link{makeWide}},
  \code{\link{niceBy}}, \code{\link{niceMerge}},
  \code{\link{path}}, \code{\link{readData}},
  \code{\link{readSurveyTool}}, \code{\link{readVector}},
  \code{\link{reportProgress}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{timeRemaining}},
  \code{\link{writeData}}, \code{\link{writeVector}}

Other sandra::ProgressTimer: \code{\link{ProgressTimer}},
  \code{\link{reportProgress}}, \code{\link{timeRemaining}}
}

