% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataFrames.R
\name{data.frame.affixNames}
\alias{data.frame.affixNames}
\title{Affix column names in a data frame}
\usage{
data.frame.affixNames(ds, prefix = NULL, postfix = NULL, not = c())
}
\arguments{
\item{ds}{(data.frame) Data frame}

\item{prefix}{(character) Prefix to add to column names (name becomes prefix.name)}

\item{postfix}{(character) Postfix to add to column names (name becomes name.postfix)}

\item{not}{(vector) Column names that should NOT be changed (get no prefix or postfix)}
}
\value{
(data.frame) Data with affixed column names
}
\description{
Affix column names in a data frame
}

