% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrames.R
\name{data.frame.dropVar}
\alias{data.frame.dropVar}
\title{Drops a variable from a data frame}
\usage{
data.frame.dropVar(ds, dropMe)
}
\arguments{
\item{ds}{(data.frame) Data frame}

\item{dropMe}{(character) Variable to drop}
}
\value{
(data.frame) ds with variable dropped
}
\description{
Drops a variable from a data frame
}
\examples{
# Create a new data frame
ds = data.frame.new( vars = c( "id", "x", "y", "z" ) );
# Drop "id" 
data.frame.dropVar( ds, "id" );
}
\seealso{
Other SANDRA: \code{\link{EasyFit}}, \code{\link{FileIO}},
  \code{\link{FrameworkFileIO}},
  \code{\link{ProgressTimer}}, \code{\link{addPostfix}},
  \code{\link{appendData}},
  \code{\link{calculateAggregation}},
  \code{\link{calculateDScores}},
  \code{\link{calculateScores}},
  \code{\link{checkAndRemoveJasminDuplicates}},
  \code{\link{cutExtension}},
  \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.new}},
  \code{\link{data.frame.renameVar}},
  \code{\link{decodeJasmin1}}, \code{\link{decodeJasmin2}},
  \code{\link{decodeSprif1}}, \code{\link{done}},
  \code{\link{dropAndReport}}, \code{\link{existsData}},
  \code{\link{getExtension}},
  \code{\link{installAnalysisFramework}},
  \code{\link{leftMerge}}, \code{\link{makeWide}},
  \code{\link{niceBy}}, \code{\link{niceMerge}},
  \code{\link{path}}, \code{\link{readData}},
  \code{\link{readSurveyTool}}, \code{\link{readVector}},
  \code{\link{reportProgress}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{timeRemaining}},
  \code{\link{writeData}}, \code{\link{writeVector}}

Other sandra::dataFrames: \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.new}},
  \code{\link{data.frame.renameVar}}
}

