% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataFrames.R
\name{data.frame.new}
\alias{data.frame.new}
\title{Creates a data frame with zero rows}
\usage{
data.frame.new(vars = c())
}
\arguments{
\item{vars}{(vector) Names of variables (columns) in data.frame}
}
\value{
(data.frame)
}
\description{
Creates a data frame with zero rows
}
\examples{
# See sandra::data.frame.affixNames
}

