% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrames.R
\name{data.frame.renameVar}
\alias{data.frame.renameVar}
\title{Renames a variable of a data frame}
\usage{
data.frame.renameVar(ds, oldName, newName)
}
\arguments{
\item{ds}{(data.frame) Data frame}

\item{oldName}{(character) Variable to rename}

\item{newName}{(character) New name of variable}
}
\value{
(data.frame) ds with variable renamed
}
\description{
Renames a variable of a data frame
}
\examples{
# Create a new data frame
ds = data.frame.new( vars = c( "id", "x", "y", "z" ) );
# Rename "id" to "foo"
data.frame.renameVar( ds, "id", "foo" );
}
\seealso{
Other SANDRA: \code{\link{EasyLM}}, \code{\link{FileIO}},
  \code{\link{FrameworkFileIO}},
  \code{\link{ProgressTimer}}, \code{\link{addPostfix}},
  \code{\link{appendData}}, \code{\link{calculateDScores}},
  \code{\link{close}}, \code{\link{cutExtension}},
  \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.new}},
  \code{\link{decodeJasmin1}}, \code{\link{existsData}},
  \code{\link{getExtension}},
  \code{\link{installAnalysisFramework}},
  \code{\link{leftMerge}}, \code{\link{makeWide}},
  \code{\link{niceBy}}, \code{\link{niceMerge}},
  \code{\link{path}}, \code{\link{readData}},
  \code{\link{readSurveyTool}}, \code{\link{readVector}},
  \code{\link{reportProgress}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{timeRemaining}},
  \code{\link{writeData}}, \code{\link{writeVector}}

Other sandra::dataFrames: \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.new}}
}

