% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodeJasmin1.R
\name{decodeJasmin1}
\alias{decodeJasmin1}
\title{Decodes JASMIN1 data stored in a LOTUS results file into trialdata and metadata}
\usage{
decodeJasmin1(input, participationID = c("UserID"), verbose = FALSE,
  timeReportInterval = 60, set_id_from = 1, colRunID = "RunID",
  colName = "Name", colValue = "Value")
}
\arguments{
\item{input}{(data.frame) LOTUS results file}

\item{participationID}{(vector) Columns in input for which each unique combination of values defines one participation}

\item{verbose}{(logical) If TRUE, then print debug output}

\item{timeReportInterval}{(integer) If 0, don't print time remaining. Else, during set conversion, print an estimate of time remaining every timeReportInterval seconds}

\item{set_id_from}{(integer) Value of set_id to start processing task data with}

\item{colRunID}{(character) Name of LOTUS RunID column}

\item{colName}{(character) Name of LOTUS Name column}

\item{colValue}{(character) Name of LOTUS Value column}
}
\value{
(list) Decoded trialdata and metadata
}
\description{
Decodes JASMIN1 data stored in a LOTUS results file into trialdata and metadata
}
\examples{
See: SANDRA/framework_demos/scripts/t.1.a Decode JASMIN1 in LOTUS.R
See: SANDRA/tutorials/2. Decoding Trial Data, Scoring Tasks, and Widening.docx
}
\seealso{
Other SANDRA: \code{\link{EasyFit}}, \code{\link{FileIO}},
  \code{\link{FrameworkFileIO}},
  \code{\link{ProgressTimer}}, \code{\link{addPostfix}},
  \code{\link{appendData}},
  \code{\link{calculateAggregation}},
  \code{\link{calculateDScores}},
  \code{\link{calculateScores}},
  \code{\link{cutExtension}},
  \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.new}},
  \code{\link{data.frame.renameVar}},
  \code{\link{decodeJasmin2}}, \code{\link{decodeSprif1}},
  \code{\link{done}}, \code{\link{dropAndReport}},
  \code{\link{existsData}}, \code{\link{getExtension}},
  \code{\link{installAnalysisFramework}},
  \code{\link{leftMerge}}, \code{\link{makeWide}},
  \code{\link{niceBy}}, \code{\link{niceMerge}},
  \code{\link{path}}, \code{\link{readData}},
  \code{\link{readSurveyTool}}, \code{\link{readVector}},
  \code{\link{reportProgress}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{timeRemaining}},
  \code{\link{writeData}}, \code{\link{writeVector}}
}

