% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodeJasmin2.R
\name{decodeJasmin2}
\alias{decodeJasmin2}
\title{Decodes JASMIN2 data stored in a LOTUS results file}
\usage{
decodeJasmin2(ds, tables, verbose = FALSE)
}
\arguments{
\item{ds}{(data.frame) LOTUS results file}

\item{verbose}{(logical) If TRUE, then print debug output}
}
\value{
(list) Decoded data
}
\description{
Decodes JASMIN2 data stored in a LOTUS results file
}
\examples{
See: SANDRA/framework_demos/scripts/t.1.c Decode JASMIN2 in LOTUS.R
See: SANDRA/tutorials/2. Decoding Trial Data, Scoring Tasks, and Widening.docx
}
\seealso{
Other SANDRA: \code{\link{EasyFit}}, \code{\link{FileIO}},
  \code{\link{FrameworkFileIO}},
  \code{\link{ProgressTimer}}, \code{\link{addPostfix}},
  \code{\link{appendData}},
  \code{\link{calculateAggregation}},
  \code{\link{calculateDScores}},
  \code{\link{calculateScores}},
  \code{\link{checkAndRemoveJasminDuplicates}},
  \code{\link{cutExtension}},
  \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.new}},
  \code{\link{data.frame.renameVar}},
  \code{\link{decodeJasmin1}},
  \code{\link{decodeJasmin2Legacy}},
  \code{\link{decodeSprif1}}, \code{\link{done}},
  \code{\link{dropAndReport}}, \code{\link{existsData}},
  \code{\link{getExtension}},
  \code{\link{installAnalysisFramework}},
  \code{\link{leftMerge}}, \code{\link{makeWide}},
  \code{\link{niceBy}}, \code{\link{niceMerge}},
  \code{\link{path}}, \code{\link{readData}},
  \code{\link{readSurveyTool}}, \code{\link{readVector}},
  \code{\link{reportProgress}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{timeRemaining}},
  \code{\link{writeData}}, \code{\link{writeVector}}
}

