% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodeSprif1.R
\name{decodeSprif1}
\alias{decodeSprif1}
\title{Decodes SPRIF data stored in a LOTUS results file into trialdata and metadata}
\usage{
decodeSprif1(ds, sprifColumn = "Value", sprifVars = NULL,
  sprifColSep = "\\t", sprifRowSep = "|", verbose = F)
}
\arguments{
\item{ds}{(data.frame) LOTUS results file}

\item{sprifColumn}{(character) Column in ds that contains SPRIF data}

\item{sprifVars}{(vector) Variables logged for each trial. In NULL, the number of variables is guessed and named "X1", "X2", etc.}

\item{sprifColSep}{(character) Character separating columns in SPRIF data}

\item{sprifRowSep}{(character) Character separating rows in SPRIF data}

\item{verbose}{(logical) If TRUE, then print debug output}
}
\value{
(list) Decoded trialdata and metadata
}
\description{
Decodes SPRIF data stored in a LOTUS results file into trialdata and metadata
}
\examples{
See: SANDRA/framework_demos/scripts/t.1.b Decode SPRIF in LOTUS.R
See: SANDRA/tutorials/2. Decoding Trial Data, Scoring Tasks, and Widening.docx
}
\seealso{
Other SANDRA: \code{\link{EasyFit}}, \code{\link{FileIO}},
  \code{\link{FrameworkFileIO}},
  \code{\link{ProgressTimer}}, \code{\link{addPostfix}},
  \code{\link{appendData}},
  \code{\link{calculateAggregation}},
  \code{\link{calculateDScores}},
  \code{\link{calculateScores}},
  \code{\link{checkAndRemoveJasminDuplicates}},
  \code{\link{cutExtension}},
  \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.new}},
  \code{\link{data.frame.renameVar}},
  \code{\link{decodeJasmin1}},
  \code{\link{decodeJasmin2Legacy}},
  \code{\link{decodeJasmin2}}, \code{\link{done}},
  \code{\link{dropAndReport}}, \code{\link{existsData}},
  \code{\link{getExtension}},
  \code{\link{installAnalysisFramework}},
  \code{\link{leftMerge}}, \code{\link{makeWide}},
  \code{\link{niceBy}}, \code{\link{niceMerge}},
  \code{\link{path}}, \code{\link{readData}},
  \code{\link{readSurveyTool}}, \code{\link{readVector}},
  \code{\link{reportProgress}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{timeRemaining}},
  \code{\link{writeData}}, \code{\link{writeVector}}
}

