% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateScores.R
\name{dropAndReport}
\alias{dropAndReport}
\title{Drops rows and reports about rows in data based on response or rt}
\usage{
dropAndReport(result, ds_subset, settings = list())
}
\arguments{
\item{result}{(list) Results so far (see ?nice_by).}

\item{ds_subset}{(data.frame) Data to apply dropping and reporting to}

\item{settings}{(list) See table below}
}
\value{
(list) Element "ds_subset" contains data with dropped responses; "result" contains responses and RTs that are reported
}
\description{
Drops rows and reports about rows in data based on response or rt
}
\details{
The table below shows an overview of elements in settings and their` meaning.
\tabular{lllll}{
  \strong{Name} \tab \strong{Type} \tab \strong{Required?} \tab  \strong{Example Value} \tab \strong{Description} \cr
  resp_drop \tab vector \tab No \tab \code{c("NA",0,3,4)} \tab Drop all responses with these values. \cr
  resp_report \tab vector \tab No \tab \code{c(1)} \tab Report number of responses with these values (per value individually) as "n_resp_<value>". \cr
  fast_drop \tab integer \tab No \tab \code{300} \tab Drop all response times lower than this value. \cr
  fast_report \tab integer \tab No \tab \code{300} \tab Report number of response times lower than this value in scores as "n_outlier_fast". This occurs after dropping rows according to resp_drop. \cr
  slow_drop \tab integer \tab No \tab \code{3000} \tab Drop all response times higher than this value. \cr
  slow_report \tab integer \tab No \tab \code{3000} \tab Report number of response times higher than this value in scores as "n_outlier_slow". This occurs after dropping rows according to resp_drop.\cr
  sd_drop \tab numeric \tab No \tab \code{1.5} \tab Drop all response times higher than mean response times of correct responses + sd_drop * SD, and those lower than mean - sd_drop * SD. \cr
  sd_report \tab numeric \tab No \tab \code{1.5} \tab Report number of response times higher than mean response times of correct responses + sd_drop * SD, or those lower than mean - sd_drop * SD in scores as "n_outlier_sd". This occurs after dropping rows according to resp_drop, slow_drop, and fast_drop. \cr
}
}
\seealso{
Other SANDRA: \code{\link{EasyFit}}, \code{\link{FileIO}},
  \code{\link{FrameworkFileIO}},
  \code{\link{ProgressTimer}}, \code{\link{addPostfix}},
  \code{\link{appendData}},
  \code{\link{calculateAggregation}},
  \code{\link{calculateDScores}},
  \code{\link{calculateScores}},
  \code{\link{cutExtension}},
  \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.new}},
  \code{\link{data.frame.renameVar}},
  \code{\link{decodeJasmin1}}, \code{\link{done}},
  \code{\link{existsData}}, \code{\link{getExtension}},
  \code{\link{installAnalysisFramework}},
  \code{\link{leftMerge}}, \code{\link{makeWide}},
  \code{\link{niceBy}}, \code{\link{niceMerge}},
  \code{\link{path}}, \code{\link{readData}},
  \code{\link{readSurveyTool}}, \code{\link{readVector}},
  \code{\link{reportProgress}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{timeRemaining}},
  \code{\link{writeData}}, \code{\link{writeVector}}
}

