% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasyLM.R
\name{EasyFit}
\alias{EasyFit}
\title{Wrapper for R's native lm function}
\usage{
EasyFit(ds, varsC = c(), varsD = c())
}
\arguments{
\item{ds}{(data.frame) Dataset to run regressions on}

\item{varsContinuous}{(vector) Converts corresponding columns in dataframe to numeric, then Z-transforms}

\item{varsDiscrete}{(vector) Converts corresponding columns in dataframe to factor}
}
\value{
(function) wrapped lm function, use as lm without specifying data argument
}
\description{
Constructs an instance of EasyLM, which adds some arguments to the native lm function for specifying which variables are continuous or discrete 
*** NOTE *** This function is still in early development
}
\seealso{
Other SANDRA: \code{\link{FileIO}},
  \code{\link{FrameworkFileIO}},
  \code{\link{ProgressTimer}}, \code{\link{addPostfix}},
  \code{\link{appendData}},
  \code{\link{calculateAggregation}},
  \code{\link{calculateDScores}},
  \code{\link{calculateScores}},
  \code{\link{checkAndRemoveJasminDuplicates}},
  \code{\link{cutExtension}},
  \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.new}},
  \code{\link{data.frame.renameVar}},
  \code{\link{decodeJasmin1}},
  \code{\link{decodeJasmin2Legacy}},
  \code{\link{decodeJasmin2}}, \code{\link{decodeSprif1}},
  \code{\link{done}}, \code{\link{dropAndReport}},
  \code{\link{existsData}}, \code{\link{getExtension}},
  \code{\link{installAnalysisFramework}},
  \code{\link{leftMerge}}, \code{\link{makeWide}},
  \code{\link{niceBy}}, \code{\link{niceMerge}},
  \code{\link{path}}, \code{\link{readData}},
  \code{\link{readSurveyTool}}, \code{\link{readVector}},
  \code{\link{reportProgress}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{timeRemaining}},
  \code{\link{writeData}}, \code{\link{writeVector}}
}

