% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EasyLM.R
\name{EasyLM}
\alias{EasyLM}
\title{Wrapper for R's native lm function}
\usage{
EasyLM(ds, varsContinuous = c(), varsDiscrete = c())
}
\arguments{
\item{ds}{(data.frame) Dataset to run regressions on}

\item{varsContinuous}{(vector) Converts corresponding columns in dataframe to numeric, then Z-transforms}

\item{varsDiscrete}{(vector) Converts corresponding columns in dataframe to factor}
}
\value{
(function) wrapped lm function, use as lm without specifying data argument
}
\description{
Constructs an instance of EasyLM, which adds some arguments to the native lm function for specifying which variables are continuous or discrete
}

