% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FileIO.R
\name{existsData}
\alias{existsData}
\title{Check if a file exists in original or interim directory}
\usage{
existsData(this, filename, original = FALSE)
}
\arguments{
\item{this}{(sandra::FileIO) FileIO instance}

\item{filename}{(character) File to check existance of}

\item{original}{(logical) If TRUE, return path to original data; if not, return path to interim data}
}
\value{
TRUE if file exists; if not, FALSE
}
\description{
Check if a file exists in original or interim directory
}
\examples{
# Create FileIO instance with folders in root of D:
io = FileIO( "D:/scripts", "D:/original", "D:/interim" );
io$existsData( "data.csv" );
}

