% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileIO.R
\name{FileIO}
\alias{FileIO}
\title{Constructs a SANDRA FileIO instance}
\usage{
FileIO(pathScripts, pathOriginal, pathInterim)
}
\arguments{
\item{pathScripts}{(character) Path to your analysis scripts}

\item{pathOriginal}{(character) Path to your original unprocessed data files}

\item{pathInterim}{(character) Path to data files as produced by your analyses}
}
\description{
Constructs a SANDRA FileIO instance
}
\examples{
# Create FileIO instance with folders in root of D:
io = FileIO( "D:/scripts", "D:/original", "D:/interim" );
}
\seealso{
Other sandra::file input & output: \code{\link{FrameworkFileIO}},
  \code{\link{addPostfix}}, \code{\link{appendData}},
  \code{\link{cutExtension}}, \code{\link{existsData}},
  \code{\link{getExtension}}, \code{\link{path}},
  \code{\link{readData}}, \code{\link{readSurveyTool}},
  \code{\link{readVector}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{writeData}},
  \code{\link{writeVector}}

Other sandra::FileIO: \code{\link{FrameworkFileIO}},
  \code{\link{appendData}}, \code{\link{existsData}},
  \code{\link{readData}}, \code{\link{readSurveyTool}},
  \code{\link{readVector}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{writeData}},
  \code{\link{writeVector}}
}

