% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FileIO.R
\name{FrameworkFileIO}
\alias{FrameworkFileIO}
\title{Constructs a SANDRA FileIO instance, assuming standard analysis, original, and interim directories}
\usage{
FrameworkFileIO(pathAnalysis)
}
\arguments{
\item{pathAnalysis}{(character) Path to your analysis}
}
\value{
(list) FileIO instance
}
\description{
Constructs a SANDRA FileIO instance, assuming standard analysis, original, and interim directories
}
\section{FileIO}{

}
\examples{
# Setup analysis for directory D:/analysis
io = FrameworkFileIO( "D:/analysis" );
}

