% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FileIO.R
\name{getExtension}
\alias{getExtension}
\title{Returns the extension of the given filename; everything after the last dot}
\usage{
getExtension(filename)
}
\arguments{
\item{filename}{(character) Filename to get extension from}
}
\value{
(character) Filename extension
}
\description{
Returns the extension of the given filename; everything after the last dot
}
\examples{
# Returns "csv"
getExtension( "data.csv" );
}

