% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAnalysisFramework.R
\name{installAnalysisFramework}
\alias{installAnalysisFramework}
\title{Installs or loads a SANDRA Analysis Framework}
\usage{
installAnalysisFramework(pathAnalysis = NA)
}
\arguments{
\item{pathAnalasis}{(character) Path to analysis directory}
}
\description{
A SANDRA Analysis Framework, which is a directory containing the following:
- Load Analysis.R Loads SANDRA and sets up a FileIO for this AnalysisFramework
- scripts/        Analysis scripts
- original/       Original Data
- interim/        Interim Data
}

