% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshaping.R
\name{makeWide}
\alias{makeWide}
\title{Wrapper for R's native reshape( direction = "wide" ) function}
\usage{
makeWide(data, idvars, timevars)
}
\arguments{
\item{data}{(data.frame) Data to reshape}

\item{idvars}{(vector) Names of one or more variables in long format that identify multiple records from the same group/individual}

\item{timevars}{(vector) Names of one or more variables in long format that differentiates multiple records from the same group or individual}
}
\value{
(data.frame) Reshaped data
}
\description{
This wrapper can handle multiple timevars (reshaped to columns dv.timevar1.timevar2)
}

