% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/niceBy.R
\name{niceBy}
\alias{niceBy}
\title{Wrapper for R's native by function}
\usage{
niceBy(ds, factors, aggregation, result_type = "data.frame", verbose = F,
  ...)
}
\arguments{
\item{ds}{(data.frame) Dataset to aggregate}

\item{factors}{(vector) Columns in dataset that serve as factors}

\item{aggregation}{(function) Aggregation}

\item{result_type}{(character) Three possible values: "data.frame", "vector", any other is interpreted as "list}

\item{verbose}{(logical) If TRUE, then print debug output}

\item{...}{Additional arguments passed to aggregation function}
}
\value{
(mixed) Aggregation result
}
\description{
Calls the function 'aggregation' for each unique combination of values of the the columns factors in the dataframe ds.
The aggregation function is called with two arguments: (1) (list) result list to add aggregation output to, and (2) (data.frame) subset of data selected by this combination of values for factors
}

