% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/merging.R
\name{niceMerge}
\alias{niceMerge}
\title{Like leftMerge but supporting left, right, and outer join}
\usage{
niceMerge(x, y, by_keys, pick_x, ...)
}
\arguments{
\item{x}{(data.frame) Left data frame}

\item{y}{(data.frame) Right data frame}

\item{by_keys}{(vector) Keys to join both data frames on}

\item{...}{Arguments passed on to R's native merge function}

\item{by_keys}{(logical) If TRUE, then for duplicates pick left table; if not, then pick right}
}
\value{
(data.frame) Joined data frames
}
\description{
Like leftMerge but supporting left, right, and outer join
}

