% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FileIO.R
\name{path}
\alias{path}
\title{Returns terms joined by "/" for constructing path names}
\usage{
path(...)
}
\arguments{
\item{...}{(character) Terms to join}
}
\value{
(character) joined terms
}
\description{
Returns terms joined by "/" for constructing path names
}
\examples{
# Returns "a/b/c"
path( "a", "b", "c" );
}

