% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrames.R
\name{rbind.easy}
\alias{rbind.easy}
\title{Wrapper for rbind}
\usage{
rbind.easy(x, y)
}
\arguments{
\item{x}{(data.frame) data frame to add to}

\item{y}{(data.frame) data frame added to x}
}
\value{
(data.frame) y added to x
}
\description{
Allows for NULL arguments and checks if names of both arguments match
}

