% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FileIO.R
\name{readData}
\alias{readData}
\title{Read a tab separated file (with varnames at the top row)}
\usage{
readData(this, filename, original = FALSE, sep = "\\t", quote = "",
  comment.char = "", header = TRUE, fill = TRUE,
  stringsAsFactors = FALSE, ...)
}
\arguments{
\item{this}{(sandra::FileIO) FileIO instance}

\item{filename}{(character) File to read}

\item{original}{(logical) If TRUE, use path to original data; if not, use path to interim data}

\item{remainder:}{Arguments passed to read.table with default values}
}
\value{
(data.frame) Dataset
}
\description{
Read a tab separated file (with varnames at the top row)
}
\examples{
# Create FileIO instance with folders in root of D:
io = FileIO( "D:/scripts", "D:/original", "D:/interim" );
io$readData( "data.csv" );
}

