% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FileIO.R
\name{readVector}
\alias{readVector}
\title{Read a vector from a newline separated file}
\usage{
readVector(this, filename, original)
}
\arguments{
\item{this}{(sandra::FileIO) FileIO instance}

\item{filename}{(character) File to read from}

\item{original}{(logical) If TRUE, use path to original data; if not, use path to interim data}
}
\value{
(vector) Vector
}
\description{
Read a vector from a newline separated file
}
\examples{
# Create FileIO instance with folders in root of D:
io = FileIO( "D:/scripts", "D:/original", "D:/interim" );
io$readVector( "data.csv" );
}

