% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProgressTimer.R
\name{reportProgress}
\alias{reportProgress}
\title{Report progress}
\usage{
reportProgress(this, stepCurrent)
}
\arguments{
\item{stepCurrent}{(integer) Number of current step}
}
\description{
If progress bar is displayed, and it is time to update it, then do so
}
\seealso{
Other SANDRA: \code{\link{EasyFit}}, \code{\link{FileIO}},
  \code{\link{FrameworkFileIO}},
  \code{\link{ProgressTimer}}, \code{\link{addPostfix}},
  \code{\link{appendData}},
  \code{\link{calculateAggregation}},
  \code{\link{calculateDScores}},
  \code{\link{calculateScores}},
  \code{\link{checkAndRemoveJasminDuplicates}},
  \code{\link{cutExtension}},
  \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.new}},
  \code{\link{data.frame.renameVar}},
  \code{\link{decodeJasmin1}}, \code{\link{decodeJasmin2}},
  \code{\link{decodeSprif1}}, \code{\link{done}},
  \code{\link{dropAndReport}}, \code{\link{existsData}},
  \code{\link{getExtension}},
  \code{\link{installAnalysisFramework}},
  \code{\link{leftMerge}}, \code{\link{makeWide}},
  \code{\link{niceBy}}, \code{\link{niceMerge}},
  \code{\link{path}}, \code{\link{readData}},
  \code{\link{readSurveyTool}}, \code{\link{readVector}},
  \code{\link{resolvePath}}, \code{\link{runScript}},
  \code{\link{timeRemaining}}, \code{\link{writeData}},
  \code{\link{writeVector}}

Other sandra::ProgressTimer: \code{\link{ProgressTimer}},
  \code{\link{done}}, \code{\link{timeRemaining}}
}

