% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FileIO.R
\name{resolvePath}
\alias{resolvePath}
\title{Returns path to original or interim data}
\usage{
resolvePath(this, original)
}
\arguments{
\item{this}{(sandra::FileIO) FileIO instance}

\item{original}{(logical) If TRUE, return path to original data; if not, return path to interim data}
}
\value{
(character) Path
}
\description{
Returns path to original or interim data
}
\examples{
# Create FileIO instance with folders in root of D:
io = FileIO( "D:/scripts", "D:/original", "D:/interim" );
# Returns "D:/original"
io$resolvePath( T );
}

