% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FileIO.R
\name{runScript}
\alias{runScript}
\title{Run a script from a file in scripts directory}
\usage{
runScript(this, script)
}
\arguments{
\item{this}{(sandra::FileIO) FileIO instance}

\item{script}{(character) File with script to run}
}
\description{
Run a script from a file in scripts directory
}
\examples{
# Create FileIO instance with folders in root of D:
io = FileIO( "D:/scripts", "D:/original", "D:/interim" );
io$runScript( "myScript.R" );
}

