% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProgressTimer.R
\name{timeSpent}
\alias{timeSpent}
\title{Report time spent (since start of ProgressTimer)}
\usage{
timeSpent(this)
}
\value{
(numeric)
}
\description{
Report time spent (since start of ProgressTimer)
}

