% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FileIO.R
\name{writeData}
\alias{writeData}
\title{Write a data frame to a tab separated file (with varnames at the top row)}
\usage{
writeData(this, filename, output, append = F, sep = "\\t", quote = F,
  row.names = F, col.names = T, ...)
}
\arguments{
\item{this}{(sandra::FileIO) FileIO instance}

\item{filename}{(character) File to write to}

\item{output}{(data.frame) Data to write to file}

\item{remainder:}{Arguments passed to read.table with default values}
}
\description{
Write a data frame to a tab separated file (with varnames at the top row)
}
\examples{
# Create FileIO instance with folders in root of D:
io = FileIO( "D:/scripts", "D:/original", "D:/interim" );
io$writeData( "data.csv", ds );
}

