% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FileIO.R
\name{writeVector}
\alias{writeVector}
\title{Write a vector to a newline separated file}
\usage{
writeVector(this, filename, output)
}
\arguments{
\item{this}{(sandra::FileIO) FileIO instance}

\item{filename}{(character) File to write to}

\item{output}{(vector) Data to write to file}
}
\description{
Write a vector to a newline separated file
}
\examples{
# Create FileIO instance with folders in root of D:
io = FileIO( "D:/scripts", "D:/original", "D:/interim" );
io$writeVector( "data.csv", v );
}

