% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FileIO.R
\name{addPostfix}
\alias{addPostfix}
\title{Postfix filename}
\usage{
addPostfix(filename, ...)
}
\arguments{
\item{filename}{(character) Filename to append postfixes to}

\item{...}{(character) Postfixes to append to filename}
}
\value{
(character) postfixed filename
}
\description{
Returns filename, with postfixes appended, each separated by dots, followed by the filename extension
}
\examples{
# Returns "data.a.b.c.csv"
addPostfix( "data.csv", "a", "b", "c" );
}
\seealso{
Other sandra::file input & output: \code{\link{FileIO}};
  \code{\link{FrameworkFileIO}}; \code{\link{appendData}};
  \code{\link{cutExtension}}; \code{\link{existsData}};
  \code{\link{getExtension}}; \code{\link{path}};
  \code{\link{readData}}; \code{\link{readSurveyTool}};
  \code{\link{readVector}}; \code{\link{resolvePath}};
  \code{\link{runScript}}; \code{\link{writeData}};
  \code{\link{writeVector}}
}

