% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FileIO.R
\name{appendData}
\alias{appendData}
\title{Create file if exists, else append data to existing file}
\usage{
appendData(this, filename, output, ...)
}
\arguments{
\item{this}{(sandra::FileIO) FileIO instance}

\item{filename}{(character) File to write to}

\item{output}{(data.frame) Data to write to file}

\item{...}{Remaining arguments passed to sandra::writeData}
}
\description{
Create file if exists, else append data to existing file
}
\examples{
# Create FileIO instance with folders in root of D:
io = FileIO( "D:/scripts", "D:/original", "D:/interim" );
io$appendData( "data.csv", ds );
}
\seealso{
Other sandra::file input & output: \code{\link{FileIO}};
  \code{\link{FrameworkFileIO}}; \code{\link{addPostfix}};
  \code{\link{cutExtension}}; \code{\link{existsData}};
  \code{\link{getExtension}}; \code{\link{path}};
  \code{\link{readData}}; \code{\link{readSurveyTool}};
  \code{\link{readVector}}; \code{\link{resolvePath}};
  \code{\link{runScript}}; \code{\link{writeData}};
  \code{\link{writeVector}}

Other sandra::FileIO: \code{\link{FileIO}};
  \code{\link{FrameworkFileIO}}; \code{\link{existsData}};
  \code{\link{readData}}; \code{\link{readSurveyTool}};
  \code{\link{readVector}}; \code{\link{resolvePath}};
  \code{\link{runScript}}; \code{\link{writeData}};
  \code{\link{writeVector}}
}

