% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FileIO.R
\name{cutExtension}
\alias{cutExtension}
\title{Returns the filename without extension; everything up to the last dot}
\usage{
cutExtension(filename)
}
\arguments{
\item{filename}{(character) Filename to cut extension from}
}
\value{
(character) Filename without extension
}
\description{
Returns the filename without extension; everything up to the last dot
}
\examples{
# Returns "data"
cutExtension( "data.csv" );
}
\seealso{
Other sandra::file input & output: \code{\link{FileIO}};
  \code{\link{FrameworkFileIO}}; \code{\link{addPostfix}};
  \code{\link{appendData}}; \code{\link{existsData}};
  \code{\link{getExtension}}; \code{\link{path}};
  \code{\link{readData}}; \code{\link{readSurveyTool}};
  \code{\link{readVector}}; \code{\link{resolvePath}};
  \code{\link{runScript}}; \code{\link{writeData}};
  \code{\link{writeVector}}
}

