% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrames.R
\name{data.frame.affixNames}
\alias{data.frame.affixNames}
\title{Affix column names in a data frame}
\usage{
data.frame.affixNames(ds, prefix = NULL, postfix = NULL, not = c())
}
\arguments{
\item{ds}{(data.frame) Data frame}

\item{prefix}{(character) Prefix to add to column names (name becomes prefix.name)}

\item{postfix}{(character) Postfix to add to column names (name becomes name.postfix)}

\item{not}{(vector) Column names that should NOT be changed (get no prefix or postfix)}
}
\value{
(data.frame) Data with affixed column names
}
\description{
Affix column names in a data frame
}
\examples{
# Create a new data frame
ds = data.frame.new( vars = c( "id", "x", "y", "z" ) );
# Prefix and postfix all names except "id"
data.frame.affixNames( ds, prefix = "pre", postfix = "post", not = "id" );
}
\seealso{
Other SANDRA: \code{\link{EasyFit}}, \code{\link{FileIO}},
  \code{\link{FrameworkFileIO}},
  \code{\link{ProgressTimer}}, \code{\link{addPostfix}},
  \code{\link{appendData}},
  \code{\link{calculateAggregation}},
  \code{\link{calculateDScores}},
  \code{\link{calculateScores}},
  \code{\link{cutExtension}},
  \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.new}},
  \code{\link{data.frame.renameVar}},
  \code{\link{decodeJasmin1}}, \code{\link{decodeSprif1}},
  \code{\link{done}}, \code{\link{dropAndReport}},
  \code{\link{existsData}}, \code{\link{getExtension}},
  \code{\link{installAnalysisFramework}},
  \code{\link{leftMerge}}, \code{\link{makeWide}},
  \code{\link{niceBy}}, \code{\link{niceMerge}},
  \code{\link{path}}, \code{\link{readData}},
  \code{\link{readSurveyTool}}, \code{\link{readVector}},
  \code{\link{reportProgress}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{timeRemaining}},
  \code{\link{writeData}}, \code{\link{writeVector}}

Other sandra::dataFrames: \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.new}},
  \code{\link{data.frame.renameVar}}
}

