% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrames.R
\name{data.frame.dropVar}
\alias{data.frame.dropVar}
\title{Drops a variable from a data frame}
\usage{
data.frame.dropVar(ds, dropMe)
}
\arguments{
\item{ds}{(data.frame) Data frame}

\item{dropMe}{(character) Variable to drop}
}
\value{
(data.frame) ds with variable dropped
}
\description{
Drops a variable from a data frame
}
\examples{
# Create a new data frame
ds = data.frame.new( vars = c( "id", "x", "y", "z" ) );
# Drop "id" 
data.frame.dropVar( ds, "id" );
}

