% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrames.R
\name{data.frame.new}
\alias{data.frame.new}
\title{Creates a data frame with zero rows}
\usage{
data.frame.new(vars = c())
}
\arguments{
\item{vars}{(vector) Names of variables (columns) in data.frame}
}
\value{
(data.frame)
}
\description{
Creates a data frame with zero rows
}
\examples{
# See sandra::data.frame.affixNames
}
\seealso{
Other SANDRA: \code{\link{EasyLM}}, \code{\link{FileIO}},
  \code{\link{FrameworkFileIO}},
  \code{\link{ProgressTimer}}, \code{\link{addPostfix}},
  \code{\link{appendData}}, \code{\link{calculateDScores}},
  \code{\link{cutExtension}},
  \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.renameVar}},
  \code{\link{decodeJasmin1}}, \code{\link{done}},
  \code{\link{existsData}}, \code{\link{getExtension}},
  \code{\link{installAnalysisFramework}},
  \code{\link{leftMerge}}, \code{\link{makeWide}},
  \code{\link{niceBy}}, \code{\link{niceMerge}},
  \code{\link{path}}, \code{\link{readData}},
  \code{\link{readSurveyTool}}, \code{\link{readVector}},
  \code{\link{reportProgress}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{timeRemaining}},
  \code{\link{writeData}}, \code{\link{writeVector}}

Other sandra::dataFrames: \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.renameVar}}
}

