% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataFrames.R
\name{data.frame.renameVar}
\alias{data.frame.renameVar}
\title{Renames a variable of a data frame}
\usage{
data.frame.renameVar(ds, oldName, newName)
}
\arguments{
\item{ds}{(data.frame) Data frame}

\item{oldName}{(character) Variable to rename}

\item{newName}{(character) New name of variable}
}
\value{
(data.frame) ds with variable renamed
}
\description{
Renames a variable of a data frame
}
\examples{
# Create a new data frame
ds = data.frame.new( vars = c( "id", "x", "y", "z" ) );
# Rename "id" to "foo"
data.frame.renameVar( ds, "id", "foo" );
}

