% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/decodeJasmin1.R
\name{decodeJasmin1}
\alias{decodeJasmin1}
\title{Decodes JASMIN1 data stored in a LOTUS results file into trial data}
\usage{
decodeJasmin1(input, participationID = c("UserID"), verbose = FALSE,
  set_id_from = 1, colRunID = "RunID", colName = "Name",
  colValue = "Value")
}
\arguments{
\item{input}{(data.frame) LOTUS results file}

\item{participationID}{(vector) Columns in input for which each unique combination of values defines one participation}

\item{verbose}{(logical) If TRUE, then print debug output}

\item{set_id_from}{(integer) Value of set_id to start processing task data with}

\item{colRunID}{(character) Name of LOTUS RunID column}

\item{colName}{(character) Name of LOTUS Name column}

\item{colValue}{(character) Name of LOTUS Value column}
}
\value{
(data.frame) Calculated d-scores
}
\description{
Decodes JASMIN1 data stored in a LOTUS results file into trial data
}
\examples{
See: SANDRA/tests/scripts/demo.jasmin1_data.2.calculateDScores.R
}

