% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FileIO.R
\name{FileIO}
\alias{FileIO}
\title{Constructs a SANDRA FileIO instance}
\usage{
FileIO(pathScripts, pathOriginal, pathInterim)
}
\arguments{
\item{pathScripts}{(character) Path to your analysis scripts}

\item{pathOriginal}{(character) Path to your original unprocessed data files}

\item{pathInterim}{(character) Path to data files as produced by your analyses}
}
\value{
(list) FileIO instance
}
\description{
Constructs a SANDRA FileIO instance
}
\section{FileIO}{

}
\examples{
# Create FileIO instance with folders in root of D:
io = FileIO( "D:/scripts", "D:/original", "D:/interim" );
}

