% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileIO.R
\name{FrameworkFileIO}
\alias{FrameworkFileIO}
\title{Constructs a SANDRA FileIO instance, assuming standard analysis, original, and interim directories}
\usage{
FrameworkFileIO(pathAnalysis, verbose = TRUE)
}
\arguments{
\item{pathAnalysis}{(character) Path to your analysis}

\item{verbose}{(logical) If TRUE, check whether each expected subdirectory exists and print that io is loaded}
}
\value{
(list) FileIO instance
}
\description{
Constructs a SANDRA FileIO instance, assuming standard analysis, original, and interim directories
}
\examples{
# Setup analysis for directory D:/analysis
io = FrameworkFileIO( "D:/analysis" );
}
\seealso{
Other sandra::file input & output: \code{\link{FileIO}},
  \code{\link{addPostfix}}, \code{\link{appendData}},
  \code{\link{cutExtension}}, \code{\link{existsData}},
  \code{\link{getExtension}}, \code{\link{path}},
  \code{\link{readData}}, \code{\link{readSurveyTool}},
  \code{\link{readVector}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{writeData}},
  \code{\link{writeVector}}

Other sandra::FileIO: \code{\link{FileIO}},
  \code{\link{appendData}}, \code{\link{existsData}},
  \code{\link{readData}}, \code{\link{readSurveyTool}},
  \code{\link{readVector}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{writeData}},
  \code{\link{writeVector}}
}

