% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileIO.R
\name{FrameworkFileIO}
\alias{FrameworkFileIO}
\title{Constructs a SANDRA FileIO instance, assuming standard analysis, original, and interim directories}
\usage{
FrameworkFileIO(pathAnalysis, verbose = TRUE)
}
\arguments{
\item{pathAnalysis}{(character) Path to your analysis}

\item{verbose}{(logical) If TRUE, check whether each expected subdirectory exists and print that io is loaded}
}
\value{
(list) FileIO instance
}
\description{
Constructs a SANDRA FileIO instance, assuming standard analysis, original, and interim directories
}
\examples{
# Setup analysis for directory D:/analysis
io = FrameworkFileIO( "D:/analysis" );
}
\seealso{
Other SANDRA: \code{\link{EasyFit}}, \code{\link{FileIO}},
  \code{\link{ProgressTimer}}, \code{\link{addPostfix}},
  \code{\link{appendData}},
  \code{\link{calculateAggregation}},
  \code{\link{calculateDScores}},
  \code{\link{calculateScores}},
  \code{\link{checkAndRemoveJasminDuplicates}},
  \code{\link{cutExtension}},
  \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.new}},
  \code{\link{data.frame.renameVar}},
  \code{\link{decodeJasmin1}},
  \code{\link{decodeJasmin2Legacy}},
  \code{\link{decodeJasmin2}}, \code{\link{decodeSprif1}},
  \code{\link{done}}, \code{\link{dropAndReport}},
  \code{\link{existsData}}, \code{\link{getExtension}},
  \code{\link{installAnalysisFramework}},
  \code{\link{leftMerge}}, \code{\link{makeWide}},
  \code{\link{niceBy}}, \code{\link{niceMerge}},
  \code{\link{path}}, \code{\link{readData}},
  \code{\link{readSurveyTool}}, \code{\link{readVector}},
  \code{\link{reportProgress}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{timeRemaining}},
  \code{\link{writeData}}, \code{\link{writeVector}}

Other sandra::file input & output: \code{\link{FileIO}},
  \code{\link{addPostfix}}, \code{\link{appendData}},
  \code{\link{cutExtension}}, \code{\link{existsData}},
  \code{\link{getExtension}}, \code{\link{path}},
  \code{\link{readData}}, \code{\link{readSurveyTool}},
  \code{\link{readVector}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{writeData}},
  \code{\link{writeVector}}

Other sandra::FileIO: \code{\link{FileIO}},
  \code{\link{appendData}}, \code{\link{existsData}},
  \code{\link{readData}}, \code{\link{readSurveyTool}},
  \code{\link{readVector}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{writeData}},
  \code{\link{writeVector}}
}

