% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAnalysisFramework.R
\name{installAnalysisFramework}
\alias{installAnalysisFramework}
\title{Installs or loads a SANDRA Analysis Framework}
\usage{
installAnalysisFramework(pathAnalysis = NA)
}
\arguments{
\item{pathAnalasis}{(character) Path to analysis directory}
}
\description{
A SANDRA Analysis Framework, which is a directory containing the following:
- Load Analysis.R Loads SANDRA and sets up a FileIO for this AnalysisFramework
- scripts/        Analysis scripts
- original/       Original Data
- interim/        Interim Data
}
\seealso{
Other SANDRA: \code{\link{EasyLM}}, \code{\link{FileIO}},
  \code{\link{FrameworkFileIO}},
  \code{\link{ProgressTimer}}, \code{\link{addPostfix}},
  \code{\link{appendData}}, \code{\link{calculateDScores}},
  \code{\link{close}}, \code{\link{cutExtension}},
  \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.new}},
  \code{\link{data.frame.renameVar}},
  \code{\link{decodeJasmin1}}, \code{\link{existsData}},
  \code{\link{getExtension}}, \code{\link{leftMerge}},
  \code{\link{makeWide}}, \code{\link{niceBy}},
  \code{\link{niceMerge}}, \code{\link{path}},
  \code{\link{readData}}, \code{\link{readSurveyTool}},
  \code{\link{readVector}}, \code{\link{reportProgress}},
  \code{\link{resolvePath}}, \code{\link{runScript}},
  \code{\link{timeRemaining}}, \code{\link{writeData}},
  \code{\link{writeVector}}
}

