% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshaping.R
\name{makeWide}
\alias{makeWide}
\title{Wrapper for R's native reshape( direction = "wide" ) function}
\usage{
makeWide(data, idvars, timevars)
}
\arguments{
\item{data}{(data.frame) Data to reshape}

\item{idvars}{(vector) Names of one or more variables in long format that identify multiple records from the same group/individual}

\item{timevars}{(vector) Names of one or more variables in long format that differentiates multiple records from the same group or individual}
}
\value{
(data.frame) Reshaped data
}
\description{
This wrapper can handle multiple timevars (reshaped to columns dv.timevar1.timevar2)
}
\seealso{
Other SANDRA: \code{\link{EasyLM}}, \code{\link{FileIO}},
  \code{\link{FrameworkFileIO}},
  \code{\link{ProgressTimer}}, \code{\link{addPostfix}},
  \code{\link{appendData}}, \code{\link{calculateDScores}},
  \code{\link{close}}, \code{\link{cutExtension}},
  \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.new}},
  \code{\link{data.frame.renameVar}},
  \code{\link{decodeJasmin1}}, \code{\link{existsData}},
  \code{\link{getExtension}},
  \code{\link{installAnalysisFramework}},
  \code{\link{leftMerge}}, \code{\link{niceBy}},
  \code{\link{niceMerge}}, \code{\link{path}},
  \code{\link{readData}}, \code{\link{readSurveyTool}},
  \code{\link{readVector}}, \code{\link{reportProgress}},
  \code{\link{resolvePath}}, \code{\link{runScript}},
  \code{\link{timeRemaining}}, \code{\link{writeData}},
  \code{\link{writeVector}}
}

