% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceBy.R
\name{niceBy}
\alias{niceBy}
\title{Wrapper for R's native by function}
\usage{
niceBy(ds, factors, aggregation, result_type = "list_to_data.frame",
  verbose = F, ...)
}
\arguments{
\item{ds}{(data.frame) Dataset to aggregate}

\item{factors}{(vector) Columns in dataset that serve as factors}

\item{aggregation}{(function) Aggregation}

\item{result_type}{(character) Three possible values: "data.frame", "vector", any other is interpreted as "list}

\item{verbose}{(logical) If TRUE, then print debug output}

\item{...}{Additional arguments passed to aggregation function}
}
\value{
(mixed) Aggregation result
}
\description{
Calls the function 'aggregation' for each unique combination of values of the the columns factors in the dataframe ds.
The aggregation function is called with two arguments: (1) (list) result list to add aggregation output to, and (2) (data.frame) subset of data selected by this combination of values for factors
If niceBy is set to return a data frame, then names of variables in this data frame 
are are drawn from the names of the elements in the list returned by the aggregation
funtion. The aggregation function can return a list of scalars if each call to 
aggregation produces one row of output, or it can return a list of list of scalars
if each call produces one or more rows of output.
}
\seealso{
Other SANDRA: \code{\link{EasyFit}}, \code{\link{FileIO}},
  \code{\link{FrameworkFileIO}},
  \code{\link{ProgressTimer}}, \code{\link{addPostfix}},
  \code{\link{appendData}},
  \code{\link{calculateAggregation}},
  \code{\link{calculateDScores}},
  \code{\link{calculateScores}},
  \code{\link{checkAndRemoveJasminDuplicates}},
  \code{\link{cutExtension}},
  \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.new}},
  \code{\link{data.frame.renameVar}},
  \code{\link{decodeJasmin1}}, \code{\link{decodeJasmin2}},
  \code{\link{decodeSprif1}}, \code{\link{done}},
  \code{\link{dropAndReport}}, \code{\link{existsData}},
  \code{\link{getExtension}},
  \code{\link{installAnalysisFramework}},
  \code{\link{leftMerge}}, \code{\link{makeWide}},
  \code{\link{niceMerge}}, \code{\link{path}},
  \code{\link{readData}}, \code{\link{readSurveyTool}},
  \code{\link{readVector}}, \code{\link{reportProgress}},
  \code{\link{resolvePath}}, \code{\link{runScript}},
  \code{\link{timeRemaining}}, \code{\link{writeData}},
  \code{\link{writeVector}}
}

