% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging.R
\name{niceMerge}
\alias{niceMerge}
\title{Like leftMerge but supporting left, right, and outer join}
\usage{
niceMerge(x, y, by_keys, pick_x, ...)
}
\arguments{
\item{x}{(data.frame) Left data frame}

\item{y}{(data.frame) Right data frame}

\item{by_keys}{(vector) Keys to join both data frames on}

\item{...}{Arguments passed on to R's native merge function}

\item{by_keys}{(logical) If TRUE, then for duplicates pick left table; if not, then pick right}
}
\value{
(data.frame) Joined data frames
}
\description{
Like leftMerge but supporting left, right, and outer join
}
\seealso{
Other SANDRA: \code{\link{EasyFit}}, \code{\link{FileIO}},
  \code{\link{FrameworkFileIO}},
  \code{\link{ProgressTimer}}, \code{\link{addPostfix}},
  \code{\link{appendData}},
  \code{\link{calculateAggregation}},
  \code{\link{calculateDScores}},
  \code{\link{calculateScores}},
  \code{\link{checkAndRemoveJasminDuplicates}},
  \code{\link{cutExtension}},
  \code{\link{data.frame.affixNames}},
  \code{\link{data.frame.dropVar}},
  \code{\link{data.frame.new}},
  \code{\link{data.frame.renameVar}},
  \code{\link{decodeJasmin1}},
  \code{\link{decodeJasmin2Legacy}},
  \code{\link{decodeJasmin2}}, \code{\link{decodeSprif1}},
  \code{\link{done}}, \code{\link{dropAndReport}},
  \code{\link{existsData}}, \code{\link{getExtension}},
  \code{\link{installAnalysisFramework}},
  \code{\link{leftMerge}}, \code{\link{makeWide}},
  \code{\link{niceBy}}, \code{\link{path}},
  \code{\link{readData}}, \code{\link{readSurveyTool}},
  \code{\link{readVector}}, \code{\link{reportProgress}},
  \code{\link{resolvePath}}, \code{\link{runScript}},
  \code{\link{timeRemaining}},
  \code{\link{updateChecksum}}, \code{\link{writeData}},
  \code{\link{writeVector}}

Other sandra::merging: \code{\link{leftMerge}}
}

