% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileIO.R
\name{path}
\alias{path}
\title{Returns terms joined by "/" for constructing path names}
\usage{
path(...)
}
\arguments{
\item{...}{(character) Terms to join}
}
\value{
(character) joined terms
}
\description{
Returns terms joined by "/" for constructing path names
}
\examples{
# Returns "a/b/c"
path( "a", "b", "c" );
}
\seealso{
Other sandra::file input & output: \code{\link{FileIO}},
  \code{\link{FrameworkFileIO}}, \code{\link{addPostfix}},
  \code{\link{appendData}}, \code{\link{cutExtension}},
  \code{\link{existsData}}, \code{\link{getExtension}},
  \code{\link{readData}}, \code{\link{readSurveyTool}},
  \code{\link{readVector}}, \code{\link{resolvePath}},
  \code{\link{runScript}}, \code{\link{writeData}},
  \code{\link{writeVector}}
}

