% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FileIO.R
\name{readSurveyTool}
\alias{readSurveyTool}
\title{Read a data.frame from a TOP Survey Tool answer file}
\usage{
readSurveyTool(this, filename, original = F)
}
\arguments{
\item{this}{(sandra::FileIO) FileIO instance}

\item{filename}{(character) File to read}

\item{original}{(logical) If TRUE, use path to original data; if not, use path to interim data}
}
\value{
(data.frame) Dataset
}
\description{
Read a data.frame from a TOP Survey Tool answer file
}
\examples{
# Create FileIO instance with folders in root of D:
io = FileIO( "D:/scripts", "D:/original", "D:/interim" );
io$readSurveyTool( "answers.csv" );
}

