% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FileIO.R
\name{writeVector}
\alias{writeVector}
\title{Write a vector to a newline separated file}
\usage{
writeVector(this, filename, output)
}
\arguments{
\item{this}{(sandra::FileIO) FileIO instance}

\item{filename}{(character) File to write to}

\item{output}{(vector) Data to write to file}
}
\description{
Write a vector to a newline separated file
}
\examples{
# Create FileIO instance with folders in root of D:
io = FileIO( "D:/scripts", "D:/original", "D:/interim" );
io$writeVector( "data.csv", v );
}
\seealso{
Other sandra::file input & output: \code{\link{FileIO}};
  \code{\link{FrameworkFileIO}}; \code{\link{addPostfix}};
  \code{\link{appendData}}; \code{\link{cutExtension}};
  \code{\link{existsData}}; \code{\link{getExtension}};
  \code{\link{path}}; \code{\link{readData}};
  \code{\link{readSurveyTool}}; \code{\link{readVector}};
  \code{\link{resolvePath}}; \code{\link{runScript}};
  \code{\link{writeData}}

Other sandra::FileIO: \code{\link{FileIO}};
  \code{\link{FrameworkFileIO}}; \code{\link{appendData}};
  \code{\link{existsData}}; \code{\link{readData}};
  \code{\link{readSurveyTool}}; \code{\link{readVector}};
  \code{\link{resolvePath}}; \code{\link{runScript}};
  \code{\link{writeData}}
}

